thread init and starting should be two separate functions
the main reason is that we need to eventually cleanup the threads if
execution is interrupted. Once the threads are initialized we need to
setup a cleanup, which should not fail if the threads are not started
yet. Once this is in place we can start the threads without race
conditions