fix recursive removal of directories
non empty directories would fail to be deleted as os.rmdir() would
refuse to remove non empty directories. Use shutl.rmtree() instead and
have a callback function to print an error in case a file / directory
fails to be removed.

Of course, being a multi threaded removal, another thread could already
have removed the file, so if the file doesn't exists already it's not an
error. This could be possibly improved as there is some overhead doing
it this way, but it's not too bad for the time being