support filenames with non UTF-8 characters by keeping original ordinal
rsync will escape all characters not valid in the current locale (see
man rsync --8-bit-output option). Output format \#oct where oct is a 3
digit octal representing the ordinal of the character

previous code was assuming utf-8 locale was used for file names and
generated file names would have different file name, despite the fact
the characters would be the same, but in different encoding, resulting
in different byte arrays for the names. For example original file name
might have t he ö letter in it with latin1 encoding, this would be
escaped by rsync running in a utf-8 environment and, when unescaped,
would result in the UTF-8 version of the ö letter. From the file system
point of view (which is locale independent, names are simply byte
arrays), these two names are different.